pyproject.toml
[tool.poetry]
name = "digital-up"
version = "0.1.0"
description = ""
authors = ["Pupupupupipu <vika566384@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
postgres = "^4.0"
sqlalchemy = "^2.0.29"
requests = "^2.31.0"
bs4 = "^0.0.2"
pydantic-settings = "^2.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
